+++ pages/ui-preivew/index.tsx
@@ -4,7 +4,7 @@ import React from 'react'
 import MainLayout from 'styles/divs/MainLayout'
 import ButtonLayout from 'styles/buttons/ButtonLayout'
 import FlexBoxLayout from 'styles/divs/FlexBoxLayout'
-import TodoListCard from '@todo/components/TodoListConatiner'
+import TodoListCard from '@todo/components/TodoListCard'
 import { doingList, doneList, todoList } from 'src/data/sample'
 
 const ButtonSamplePage: NextPage = () => {
